cleanup: Silence compilation warnings on MinGW-w64
Some of the code, especially the dshow and windows codec loader parts,
are extremely hacky and likely full of bugs. The goal is merely getting
rid of warnings that could obscure more important warnings and actual
bugs, instead of fixing actual problems. This reduces the number of
warnings from over 500 to almost the same as when compiling on Linux.

Note that many problems stem from using the ancient wine-derived
windows headers. There are some differences to the "proper" windows
header. Changing the code to compile with the proper headers would be
too much trouble, and it still has to work on Unix.

Some of the changes might actually break compilation on legacy MinGW,
but we don't support that anymore. Always use MinGW-w64, even when
compiling to 32 bit.

Fixes some warnings in the win32 loader code on Linux too.

+++ osdep/getch2-win.c
@@ -77,7 +77,7 @@ static int getch2_internal(void)
     /*check if there are input events*/
 	if(!GetNumberOfConsoleInputEvents(in,&retval))
 	{
-		printf("getch2: can't get number of input events: %i\n",GetLastError());
+		printf("getch2: can't get number of input events: %i\n",(int)GetLastError());
 		return -1;
 	}
     if(retval<=0)return -1;
@@ -165,7 +165,7 @@ void getch2_enable(void)
     in = GetStdHandle(STD_INPUT_HANDLE);
    	if(!GetNumberOfConsoleInputEvents(in,&retval))
 	{
-		printf("getch2: %i can't get number of input events  [disabling console input]\n",GetLastError());
+		printf("getch2: %i can't get number of input events  [disabling console input]\n",(int)GetLastError());
 		getch2_status = 0;
 	}
     else getch2_status=1;
